Extract Redirect Behavior to own module
